rework vmmap check

+++ src/trusted/service_runtime/nacl_syscall_common.c
@@ -4251,7 +4251,7 @@ int32_t NaClSysExecve(struct NaClAppThread *natp, char const *path, char *const 
   /* Make sys_envp_ptr a NULL array if we were passed NULL by EXECV */
   if (envp) {
     sys_envp_ptr = (uint32_t*)NaClUserToSysAddrProt(nap, (uintptr_t)envp, NACL_ABI_PROT_READ);
-    int not_byte_aligned = sys_envp_ptr % 8;
+    int not_byte_aligned = (uintptr_t)sys_envp_ptr % 8;
     if (not_byte_aligned || (kNaClBadAddress == sys_envp_ptr)) {
       NaClLog(2, "NaClSysExecve could not translate environment address, returning %d\n", -NACL_ABI_EFAULT);
       ret = -NACL_ABI_EFAULT;
@@ -4340,7 +4340,7 @@ int32_t NaClSysExecv(struct NaClAppThread *natp, char const *path, char *const *
   */
   sys_argv_ptr = (uint32_t*)NaClUserToSysAddrProt(nap, (uintptr_t)argv, NACL_ABI_PROT_READ);
 
-  int not_byte_aligned = sys_argv_ptr % 8;
+  int not_byte_aligned = (uintptr_t)sys_argv_ptr % 8;
   if (not_byte_aligned || (kNaClBadAddress == sys_argv_ptr)) {
     NaClLog(2, "NaClSysExecv could not translate argv address, returning %d\n", -NACL_ABI_EFAULT);
     ret = -NACL_ABI_EFAULT;
